requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
jinja2==3.1.4
python-multipart==0.0.9
pillow==10.4.0
pydantic==2.9.1